+++ tests/epyccel/test_epyccel_sets.py
@@ -203,3 +203,83 @@ def test_Discard_wrong_arg(language):
     pyccel_result = epyccel_remove()
     python_result = Discard_wrong_arg()
     assert python_result == pyccel_result
+
+def test_update_basic(language):
+    def update_basic():
+        a = {1, 2, 3}
+        b = {4, 5, 6}
+        a.update(b)
+        return a
+
+    epyccel_update = epyccel(update_basic, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_basic()
+    assert python_result == pyccel_result
+
+def test_update_multiple(language):
+    def update_multiple():
+        a = {1, 2, 3}
+        a.update({4, 5})
+        a.update({6, 7, 8, 9})
+        a.update({10})
+        return a
+
+    epyccel_update = epyccel(update_multiple, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_multiple()
+    assert python_result == pyccel_result
+
+
+def test_update_boolean_tuple(language):
+    def update_boolean_tuple():
+        a = {True}
+        b = (False, True, False)
+        a.update(b)
+        return a
+    epyccel_update = epyccel(update_boolean_tuple, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_boolean_tuple()
+    assert python_result == pyccel_result
+
+
+def test_update_complex_list(language):
+    def update_complex_list():
+        a = {1j, 2 + 3j, 0 + 0j}
+        b = {4j, 5j, 1 + 6j}
+        a.update(b)
+        return a
+    epyccel_update = epyccel(update_complex_list, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_complex_list()
+    assert python_result == pyccel_result
+
+def test_update_range(language):
+    def update_range():
+        a = {1, 2, 3}
+        a.update(range(4, 9))
+        return a
+    epyccel_update = epyccel(update_range, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_range()
+    assert python_result == pyccel_result
+
+def test_update_set_as_arg(language):
+    def update_set_as_arg():
+        a = {1, 2, 3}
+        a.update({4, 5, 6})
+        return a
+
+    epyccel_update = epyccel(update_set_as_arg, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_set_as_arg()
+    assert python_result == pyccel_result
+
+def test_update_tuple_as_arg(language):
+    def update_tuple_as_arg():
+        a = {1, 2, 3}
+        a.update((4, 5, 6))
+        return a
+    epyccel_update = epyccel(update_tuple_as_arg, language=language)
+    pyccel_result = epyccel_update()
+    python_result =  update_tuple_as_arg()
+    assert python_result == pyccel_result